refactor: move types to one file
