+++ src/components/card/Card.tsx
@@ -1,7 +1,9 @@
 import { Component } from 'react';
-import { CardPropsType } from '../../../../types/CardPropsType';
+
+import { CardPropsType } from '../../types';
 import NoImage from './no-image.png';
 import styles from './Card.module.css';
+
 
 export class Card extends Component<CardPropsType> {
   constructor(props: CardPropsType) {
