--- src/types/AppStateType.ts
@@ -1,8 +1,0 @@
-import { CharacterType } from './CharacterType';
-
-export type AppStateType = {
-  charactersList: CharacterType[];
-  isLoading: boolean;
-  error: string;
-  searchTerm: string;
-};
