--- src/types/CardListPropsType.ts
@@ -1,5 +1,0 @@
-import { CharacterType } from "./CharacterType";
-
-export type CardListPropsType = {
-  charactersList: CharacterType[];
-}
