--- src/types/CardPropsType.ts
@@ -1,5 +1,0 @@
-import { CharacterType } from './CharacterType';
-
-export type CardPropsType = {
-  character: CharacterType;
-};
